Leksion 11 impro 6

+++ part_07/exercise_02.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 class Person:
     name: str
     citizenship: str
-    worth: int
+    worth: float
 
 
 def import_persons(path: str) -> list[Person]:
